# Conflicts:
#	scm-common/src/main/java/ir/daneshrefah/scm/common/model/asset/Membership.java
#	scm-common/src/main/java/ir/daneshrefah/scm/common/model/asset/MembershipTerminalAccess.java
#	scm-common/src/main/java/ir/daneshrefah/scm/common/model/customer/UserProfile.java
#	scm-core/src/main/java/ir/daneshrefah/scm/core/authority/decision/voter/AssetAssignmentDecisionVoter.java
#	scm-core/src/main/java/ir/daneshrefah/scm/core/entity/asset/MembershipEntity.java
#	scm-core/src/main/java/ir/daneshrefah/scm/core/entity/asset/MembershipTerminalAccessEntity.java
#	scm-core/src/main/java/ir/daneshrefah/scm/core/integration/provider/CustomerServiceImpl.java
#	scm-core/src/main/java/ir/daneshrefah/scm/core/mapper/MembershipMapper.java
#	scm-core/src/main/java/ir/daneshrefah/scm/core/repository/MembershipRepository.java
#	scm-core/src/main/java/ir/daneshrefah/scm/core/transformer/AccountListResponseTransformer.java
#	scm-plugin-api/src/main/java/ir/daneshrefah/scm/plugin/api/expose/CustomerManagementService.java
#	scm-plugin-api/src/main/java/ir/daneshrefah/scm/plugin/api/service/CustomerService.java
#	scm-uaa/src/main/java/ir/daneshrefah/scm/uaa/repository/activation/UserActivationEntity.java
#	scm-uaa/src/main/java/ir/daneshrefah/scm/uaa/service/otp/provder/SmsOtpProvider.java
#	scm-uaa/src/main/java/ir/daneshrefah/scm/uaa/service/user/UserService.java
#	scm-utils/src/main/java/ir/daneshrefah/scm/utils/string/StringUtils.java

+++ scm-common-data/src/main/java/ir/daneshrefah/scm/common/data/service/person/AbstractPersonServiceDatabaseImpl.java
@@ -73,7 +73,7 @@ public abstract class AbstractPersonServiceDatabaseImpl implements PersonService
     }
 
     @Override
-    public GeneralPerson findPersonByPersonId(Integer id) {
+    public GeneralPerson findPersonByPersonId(Long id) {
         ValidationUtils.checkNumericInput(String.valueOf(id), () -> new InvalidInputException("id"));
         Optional<GeneralPersonEntity> personEntity = personRepository.findById(id);
         ValidationUtils.checkEmptyOptional(personEntity,()->new PersonNotFoundException("person with id '" + id + "' not found."
