+++ scm-task/src/main/java/ir/daneshrefah/scm/task/service/TaskManagementServiceImpl.java
@@ -206,7 +206,7 @@ public class TaskManagementServiceImpl implements TaskManagementService {
 
     public List<TaskResponse> findAllTasksByProcessId(Long processID) {
         ProcessInstanceEntity processInstance = processManagementService.findByID(processID);
-        Integer loggedInUserId = AuthenticationUtils.getLoggedInUserId();
+        Long loggedInUserId = AuthenticationUtils.getLoggedInUserId();
         boolean hasAccess = processInstance.getTasks()
                 .stream()
                 .anyMatch(task -> loggedInUserId.equals(task.getUserId())) || loggedInUserId.equals(processInstance.getConfirmUserId());
